% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{CalculateSCInputs}
\alias{CalculateSCInputs}
\title{Calculate SCInput values on the base of SC raw data}
\usage{
CalculateSCInputs(
  state,
  ids,
  flagSetEvent = TRUE,
  flagSynchronous = getOption("MMVshiny.synchSetSCInput", FALSE)
)
}
\arguments{
\item{state}{a state object}

\item{ids}{character vector containing the ids of input parameters for which the
values have to be updated.}

\item{flagSetEvent}{logical (default: TRUE) indicating if an SCDATA event should be
fired. Currently, the purpose of this argument is to prevent firing SCDATA events
when loading Science Cloud raw data from a MMVSola report file.}

\item{flagSynchronous}{logical indicating if the SCDATA event should be
processed within this function call (TRUE) or if it should be processed when an observer
on GetSCInput(id) is triggered (FALSE). By default this is set to \code{getOption("MMVshiny.synchSetSCInput", FALSE)}.
Pay attention that setting this argument to TRUE may lead to unexpected behavior if there
is an observer on GetSCInput(id). This argument is passed to SetSCInput.}
}
\value{
nothing - this function only has a side-effect, namely calling SetSCInput and SetEvent("SCDATA").
}
\description{
Applies the SCFILTER and SCVALUE expressions from the input specification in the context
of raw SC data uploaded by the user or changed after a data curation event (the user checking
or unchecking a row).
}
