% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{cat2}
\alias{cat2}
\title{A function which cats depending on \code{getOption("MMVshiny.verbose", FALSE)}}
\usage{
cat2(...)
}
\arguments{
\item{...}{Code which is passed to \code{\link{cat}}}
}
\value{
Nothing, the result is a side-effect: Printing messages
}
\description{
A function which cats depending on \code{getOption("MMVshiny.verbose", FALSE)}
}
