% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{Get}
\alias{Get}
\title{Get validated value stripped of attributes for id in a state object}
\usage{
Get(state, id)
}
\arguments{
\item{state}{a state object.}

\item{id}{a character string id.}
}
\value{
the currently validated value for id in state.
}
\description{
Get validated value stripped of attributes for id in a state object
}
