% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetDecDigits}
\alias{GetDecDigits}
\title{For a numeric input, get the DECDIGITS value to use for rounding}
\usage{
GetDecDigits(state, id, spec = state$spec)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a numeric input parameter.}

\item{spec}{the parameter specification table. Default: \code{state$spec}.}
}
\value{
if id is a numeric input, the value of the DECDIGITS column for it, otherwise NA.
}
\description{
For a numeric input, get the DECDIGITS value to use for rounding
}
