% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetDefault}
\alias{GetDefault}
\title{Evaluate the default expression for an input parameter in a state}
\usage{
GetDefault(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the result of evaluating the default expression for id in state. If the
default expression evaluates to NULL \code{NAVal(state, id)} is returned.
}
\description{
Evaluate the default expression for an input parameter in a state
}
\details{
This function is not supposed to be called by end users. It is used by
automatically generated observer code. Since this function evaluates a reactive
expression it can only be called within a reactive context.
}
