% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetDisplayed}
\alias{GetDisplayed}
\title{Get the displayed value in a GUI input field for an input parameter id in a state object.}
\usage{
GetDisplayed(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the displayed for id.
}
\description{
Get the displayed value in a GUI input field for an input parameter id in a state object.
}
