% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetEvent}
\alias{GetEvent}
\title{Get the current event for an input parameter id in a state object}
\usage{
GetEvent(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
a character string, one of "USER", "SCDATA", "REPORT", "INIT", "DEFAULT", "MIN", "MAX".
}
\description{
Get the current event for an input parameter id in a state object
}
\details{
This function is called from Process...Event functions.
}
