% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetGuiId}
\alias{GetGuiId}
\title{Get the id of the GUI input element associated with parameter id.}
\usage{
GetGuiId(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
a character string.
}
\description{
The GUI input element id is the concatenation of state$stateId, '_', and id,
unless a prefixGuiId is specified in state, in which case the GUI input element id is:
\itemize{
\item id if state$prefixGuiId == ''
\item the concatenation of state$prefixGuiId, '_', and id, otherwise.
}
}
