% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetGuiInput}
\alias{GetGuiInput}
\title{Get user input value for input parameter id in state.}
\usage{
GetGuiInput(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the guiInput value for id in state.
}
\description{
Get user input value for input parameter id in state.
}
