% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetGuiLabel}
\alias{GetGuiLabel}
\title{Get the GUILABEL for one or several ID's in the spec}
\usage{
GetGuiLabel(state, ids, spec = state$spec)
}
\arguments{
\item{ids}{a character vector indicating ids for which the GUILABEL should be returned}
}
\value{
a character vector
}
\description{
Get the GUILABEL for one or several ID's in the spec
}
