% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetMin}
\alias{GetMin}
\title{Get the min value for a numeric input id in a state}
\usage{
GetMin(state, id, spec = state$spec)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a numeric input parameter.}

\item{spec}{the parameter specification table. Default: \code{state$spec}.}
}
\value{
if id is a numeric input, the value of the MIN expression for it, otherwise an error is thrown.
}
\description{
Get the min value for a numeric input id in a state
}
