% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetMinNote}
\alias{GetMinNote}
\title{Get the reactive note associated with the min value of a numeric input in a state}
\usage{
GetMinNote(state, id, spec = state$spec)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a numeric input parameter.}

\item{spec}{the parameter specification table. Default: \code{state$spec}.}
}
\value{
if id is a numeric input, the value of the MINNOTE expression for it, otherwise NA.
}
\description{
Get the reactive note associated with the min value of a numeric input in a state
}
