% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetOutOfBounds}
\alias{GetOutOfBounds}
\title{Get the out-of-bound policy for a numeric input id in a state object}
\usage{
GetOutOfBounds(state, id)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a numeric input parameter.}
}
\value{
a character string, one of "nearest", or "na", or NA_character_ if id is not a
numeric input.
}
\description{
Get the out-of-bound policy for a numeric input id in a state object
}
