% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetRadioChoices}
\alias{GetRadioChoices}
\title{Get the possible choices for a radio input from the state specification}
\usage{
GetRadioChoices(state, id, spec = state$spec)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the radio input parameter.}

\item{spec}{the parameter specification table. Default: \code{state$spec}.}
}
\value{
a character vector of the possible choices for the radio input.
}
\description{
Get the possible choices for a radio input from the state specification
}
