% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetReportInput}
\alias{GetReportInput}
\title{Get the value for an id inputted from a report}
\usage{
GetReportInput(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the reportInput value for id in state.
}
\description{
Get the value for an id inputted from a report
}
\details{
This function is not supposed to be called by end users. Given that
the reportInput is a reactiveValue, this function can only be called from
within a reactive context.
}
