% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetResetCount}
\alias{GetResetCount}
\title{Get the reset count for an input parameter id in state}
\usage{
GetResetCount(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the reset count for id in state.
}
\description{
Get the reset count for an input parameter id in state
}
\details{
This function is not supposed to be called by end users but from
automatically generated observers. Since the resetCount is a reactiveValue,
this function can only be called from within a reactive context.
}
