% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetSCInput}
\alias{GetSCInput}
\title{Get the value for an id averaged from an uploaded Science Cloud file}
\usage{
GetSCInput(state, id)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of the input parameter}
}
\value{
the input value
}
\description{
Get the value for an id averaged from an uploaded Science Cloud file
}
