% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetSCRawData}
\alias{GetSCRawData}
\title{Get the Science Cloud raw data in a state object}
\usage{
GetSCRawData(state, copy = FALSE)
}
\arguments{
\item{state}{a state object.}

\item{copy}{logical (default FALSE), indicating whether a copy of the reactiveVal data.table should be returned
Set to TRUE if you intend to manipulate the SCRawData for displaying purpose, e.g. change column-names to
wrappable strings, etc.}
}
\value{
a data.table or NULL, if no Science Cloud data has been uploaded
}
\description{
Get the Science Cloud raw data in a state object
}
