% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetSource}
\alias{GetSource}
\title{Get source for validated value for id in a state object}
\usage{
GetSource(state, id)
}
\arguments{
\item{state}{a state object.}

\item{id}{a character string id.}
}
\value{
a character string: For inputs, one of "User input", "Report input", "Default value";
For numeric constants the value for id in the SOURCE column of state$spec.
}
\description{
Get source for validated value for id in a state object
}
