% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetStateId}
\alias{GetStateId}
\title{Get the state identifier for a state object}
\usage{
GetStateId(state)
}
\arguments{
\item{state}{a state object.}
}
\value{
a character string representing the state identifier - this is the stateId argument passed to InitState,
or a unique identifier generated by NextStateId if "\if{html}{\out{<auto>}}" was passed as the stateId argument.
}
\description{
Get the state identifier for a state object
}
