% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetStatus}
\alias{GetStatus}
\title{Get the status for an input parameter id in a state object}
\usage{
GetStatus(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
a list with the status for id in state.
}
\description{
Get the status for an input parameter id in a state object
}
