% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetType}
\alias{GetType}
\title{Get the type of a parameter in a state object}
\usage{
GetType(state, id, spec = state$spec)
}
\arguments{
\item{state}{the state object}

\item{id}{the id of the parameter}

\item{spec}{the parameter specification table. Default: \code{state$spec}. This argument
only needs to be specified if calling outside of a reactive context without
initialized state.

Supported parameter types are "text input", "numeric input", "radio input", "numeric constant"}
}
\value{
the TYPE column value for id in the parameter spec
}
\description{
Get the type of a parameter in a state object
}
