% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetUnit}
\alias{GetUnit}
\title{Get the unit of a parameter in a state object}
\usage{
GetUnit(state, id, spec = state$spec)
}
\arguments{
\item{state}{the state object}

\item{id}{the id of the parameter}

\item{spec}{the parameter specification table. Default: \code{state$spec}. This argument
only needs to be specified if calling outside of a reactive context without
initialized state.}
}
\value{
the UNIT column value for id in the parameter spec
}
\description{
Get the unit of a parameter in a state object
}
