% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetValidated}
\alias{GetValidated}
\title{Get the validated value of an input parameter id in a state object}
\usage{
GetValidated(state, id)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}
}
\value{
the validate value for id in state.
}
\description{
Get the validated value of an input parameter id in a state object
}
