% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetValidationNote}
\alias{GetValidationNote}
\title{Get validation note for validated value for id in a state object}
\usage{
GetValidationNote(state, id, spec = state$spec)
}
\arguments{
\item{state}{a state object.}

\item{id}{a character string id.}
}
\value{
a character vector.
}
\description{
Get validation note for validated value for id in a state object
}
