% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetValidRangeNote}
\alias{GetValidRangeNote}
\title{Get the valid range note for a numeric input}
\usage{
GetValidRangeNote(state, id, spec = state$spec)
}
\arguments{
\item{state}{the state}

\item{id}{the id of the numeric input}

\item{spec}{the spec. Default is state$spec.}
}
\value{
a character string with the valid range note.
}
\description{
Get the valid range note for a numeric input
}
