% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{InitState}
\alias{InitState}
\title{Initialize state object based on a parameter specification}
\usage{
InitState(
  spec,
  stateId = "<auto>",
  listObjects = NULL,
  FLAG_ignoreNameConflicts = FALSE,
  scriptsToSource = NULL
)
}
\arguments{
\item{spec}{a data.table denoting the parameter specification.}

\item{stateId}{a character string to be used as an identifier of this state object in lists of states.
If this is set to "<auto>" (default), the id will be set to "stateX" where X is a unique number
for the duration of the session. Unless listObjects defines a prefixGuiId element, the stateId will
be used as a prefix of GUI inputs for all parameters in the state (see also \code{\link{GetGuiId}}).}

\item{listObjects}{a named list of objects to be added to the state environment. No checks are done for conflicts
with the names created in the state during initialization (see the full list in the Value section).}

\item{scriptsToSource}{a character vector of script file paths to source in the state environment.}
}
\value{
an initialized state environment containing the following objects (created in this order):
\itemize{
\item MYSTATE: A reference to the state object itself
\item stateId: The stateId argument or automatically generated unique identifier if "<auto>" was passed as the stateId.
\item spec: The spec argument
\item status: A reactiveValues object containing the status of each input parameter
\item statusIconClass: A reactiveValues object containing the icon class for the status of each input parameter
\item statusIconColor: A reactiveValues object containing the icon color for the status of each input parameter
\item statusTitle: A reactiveValues object containing the title for the status of each input parameter
\item statusText: A reactiveValues object containing the text for the status of each input parameter
\item event: A reactiveValues object containing the latest event for each input parameter
\item validated: A reactiveValues object containing the validated value for each input parameter
\item displayed: A reactiveValues object containing the displayed value for each input parameter
\item resetCount: A reactiveValues object containing the reset count for each input parameter - this
is incremented when the Reset button is pressed to trigger observers.
\item guiInput: A reactiveValues object containing the value inputted by the user in the gui input field
for each input parameter
\item reportInput: A reactiveValues object containing the value read from the Basic Inputs or Advanced Inputs
tab of a MMVSola report for each input parameter
\item scInput: A reactiveValues object containing the average value from the included records matching the SCFILTER column
of each input parameter
\item scRawData: A reactiveVal object containing an inputted Science Cloud extract set by SetSCRawData
\item scRawDataChanged: A reactiveVal object containing a flag to indicate that the SC data has changed
\item default: A list object containing the default value or parsed expression for each input parameter read from the VALEXPR column
of spec
\item defaultNote: A list object containing the default value note or parsed expression for each input parameter read from the
DEFAULTVALUENOTE column of spec
\item min: A list object containing the min value or parsed expression for each numeric input parameter read from the MIN column of spec
\item minNote: A list object containing the min value note or parsed expression for each numeric input parameter read from the MINNOTE
column of spec
\item max: A list object containing the max value or parsed expression for each numeric input parameter read from the MAX column of spec
\item maxNote: A list object containing the max value note or parsed expression for each numeric input parameter read from the MAXNOTE
column of spec
\item objects created from the listObjects argument
\item objects created from sourcing the R scripts in the scriptsToSource argument
}
}
\description{
Initialize state object based on a parameter specification
}
