% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{isSCRawData}
\alias{isSCRawData}
\title{Check if a data.frame is a Science Cloud format}
\usage{
isSCRawData(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
TRUE or FALSE
}
\description{
Check if a data.frame is a Science Cloud format
}
