% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{LoadStateSpecification}
\alias{LoadStateSpecification}
\title{Read input specification file and add columns with reactive expressions
for MIN, MAX, DEFAULT, MINNOTE, MAXNOTE and DEFNOTE.}
\usage{
LoadStateSpecification(filename)
}
\arguments{
\item{filename}{a character string indicating an xlsx file specifying input parameter
specification. It is expected that the file contains at least the following columns:
\itemize{
\item ID: This column is used to identify each input parameter in the spec.
\item TYPE: a character string column defining the type of the input.
\item VALEXPR: This column is used in the creation of the ExprVAL reactive
expression column.
\item DEFAULTVALUENOTE: This column is used in the creation of the ExprDEFNOTE
reactive expression column. It is passed to the ParseValExpr function.
\item MIN: This column is used in the creation of the ExprMIN reactive expression
column, but only for rows where the type is "numeric input".
\item MINNOTE: This column is used in the creation of the ExprMINNOTE reactive
expression column, but only for rows where the type is "numeric input".
\item MAX: This column is used in the creation of the ExprMAX reactive expression
column, but only for rows where the type is "numeric input".
\item MAXNOTE: This column is used in the creation of the ExprMAXNOTE reactive
expression column, but only for rows where the type is "numeric input".
}}
}
\value{
a data.table representing the input specification data with added columns
ExprVAL, ExprDEFNOTE, ExprMIN, ExprMINNOTE, ExprMAX, and ExprMAXNOTE. An
error is thrown if some of the IDs are duplicated.
}
\description{
The function looks for expressions of the form Get(something) and replaces
them with state\%>\%Get('something').
The function also checks if all the IDs in the Get expressions are present in the
ID column of the spec data table. If not, it stops execution and throws an error.
}
\details{
Depending on the type of the ID, it then creates reactive expression columns in
the spec data table. These reactive expressions are created by wrapping the expressions
in reactive({}) and assigning them to new columns in the spec data table. The names
of these new columns are ExprVAL, ExprDEFNOTE, ExprMIN, ExprMINNOTE, ExprMAX,
and ExprMAXNOTE.
}
