% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{NAVal}
\alias{NAVal}
\title{Get the NA value of the correct type corresponding to id}
\usage{
NAVal(state, id, spec = state$spec, type = GetType(id = id, spec = spec))
}
\arguments{
\item{state}{the state object}

\item{id}{the id of the input}

\item{spec}{the spec object. Default is state$spec.}

\item{type}{character string, the TYPE of id in spec. Default is \code{GetType(id = id, spec = spec)}.
This argument only needs to be specified when the caller has already looked up the type, to avoid
re-deriving it from spec.}
}
\value{
the NA value of the correct type
}
\description{
Get the NA value of the correct type corresponding to id
}
