% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{NextStateId}
\alias{NextStateId}
\title{Generate next unique prefix for GUI parameter inputs}
\usage{
NextStateId()
}
\value{
a character string representing the next unique prefix for GUI parameter inputs
in the formate stateX where X is a number.
}
\description{
Generate next unique prefix for GUI parameter inputs
}
