% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessDefaultChangedEvent}
\alias{ProcessDefaultChangedEvent}
\title{Validate default input value after it has changed}
\usage{
ProcessDefaultChangedEvent(s, id)
}
\arguments{
\item{s}{the state object.}

\item{id}{the ID of the input parameter.}
}
\value{
the validated value for the input or nothing if not
\code{is.na(s\%>\%GetEvent(id)) && s\%>\%GetSource(id) == "Default value"}.
}
\description{
Validate default input value after it has changed
}
\details{
This function is supposed to be called from an observer of GetDefault(s, id).
The call to this function will not do anything if not
\code{is.na(s\%>\%GetEvent(id)) && s\%>\%GetSource(id) == "Default value"}.
}
