% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessGuiInputEvent}
\alias{ProcessGuiInputEvent}
\title{Update the state when the user changes the value in a GUI input field}
\usage{
ProcessGuiInputEvent(s, input, id, inputValue)
}
\arguments{
\item{s}{the state object}

\item{input}{the shiny input object.}

\item{id}{the ID of the input field}

\item{inputValue}{the value of the input field. If this parameter is left missing,
the value of the input field is read from the shiny input object. Otherwise, the
value of the input field from the shiny input object is ignored and this value is
used instead.}
}
\value{
the validated input value based on the user input
}
\description{
Update the state when the user changes the value in a GUI input field
}
