% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessMaxChangedEvent}
\alias{ProcessMaxChangedEvent}
\title{Validate the current value of an input parameter after its MAX has changed and
update the info icon}
\usage{
ProcessMaxChangedEvent(s, id)
}
\arguments{
\item{s}{the state object.}

\item{id}{the ID of the input parameter.}
}
\value{
the validated value for the input.
}
\description{
Validate the current value of an input parameter after its MAX has changed and
update the info icon
}
\details{
This function is supposed to be called from an observer of GetMax(s, id). The
MAX value for an input parameter can changed if it is dependent on another input.
}
