% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessReportInputEvent}
\alias{ProcessReportInputEvent}
\title{Validate an input value read from a report, update the displayed value and the
info icon}
\usage{
ProcessReportInputEvent(s, id)
}
\arguments{
\item{s}{the state object.}

\item{id}{the ID of the input parameter.}
}
\value{
the validated input value or nothing if not
\code{isTRUE(GetEvent(s, id) == "REPORT")}.
}
\description{
Validate an input value read from a report, update the displayed value and the
info icon
}
\details{
Calling this function will not do anything if not \code{isTRUE(GetEvent(s, id) == "REPORT")}.
This function is not supposed to be called by users, but by automatically generated reactive observers.
}
\seealso{
\code{\link{GenerateScriptCreatingObservers}}
}
