% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessResetEvent}
\alias{ProcessResetEvent}
\title{Re-initialize an input after the user has invoked reset for the state object to which it belongs}
\usage{
ProcessResetEvent(s, id)
}
\arguments{
\item{s}{the state object}

\item{id}{the ID of the input field}
}
\value{
the validated re-initialized value for the input
}
\description{
Re-initialize an input after the user has invoked reset for the state object to which it belongs
}
\details{
This function is supposed to be called from an observer of GetResetCount(s, id).
Calling this function will re-initialize the input field to its default value and update its
displayed value, color and info icon.
}
