% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ProcessSCInputEvent}
\alias{ProcessSCInputEvent}
\title{Validate a new SC input value, update the displayed value and the info icon}
\usage{
ProcessSCInputEvent(s, id)
}
\arguments{
\item{s}{the state object.}

\item{id}{the ID of the input parameter.}
}
\value{
the validated value for the input.
}
\description{
Validate a new SC input value, update the displayed value and the info icon
}
\details{
This function is supposed to be called from an observer of GetSCInput(s, id).
The call to this function will not do anything if not \code{isTRUE(GetEvent(s, id) == "SCDATA")}.
This function is not supposed to be called by users, but by automatically generated reactive observers.
}
\seealso{
\code{\link{SetSCInput}}, \code{\link{GenerateScriptCreatingObservers}}
}
