% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ReadSCInput}
\alias{ReadSCInput}
\title{Read value from a Science Cloud input file}
\usage{
ReadSCInput(dataSC, filter, exprValue, NAvalue = NA_real_, exprFilter = NULL)
}
\arguments{
\item{dataSC}{a data.frame representing the Science Cloud input file. Can be NULL, in which case NAvalue is returned.}

\item{filter}{a named list with names corresponding to columns in dataSC and values
representing filter expressions - in these expressions the '$$' is treated as a blank
and is replaced by the columnName. After blank replacement, the filter elements are
combined into one expression using &, i.e. logical AND. If a non NULL exprFilter is specified,
this argument is ignored.}

\item{exprValue}{an R expression evaluated within the environment of the filtered data to calculate the final value.}

\item{NAvalue}{a value to return in case of an error or if the filtering resulted in empty data.}

\item{exprFilter}{NULL (default) or character string denoting a filter expression. If not NULL, the filter
argument will be ignored.}
}
\value{
NAvalue in case of empty filtered data or an error during filtering or value evaluation. Otherwise, the result
of evaluating exprValue on the filtered data.
}
\description{
Read value from a Science Cloud input file
}
