% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{Reset}
\alias{Reset}
\title{Reinitialize a state object}
\usage{
Reset(state)
}
\arguments{
\item{state}{the state object.}
}
\value{
nothing - calling this function only has side effects.
}
\description{
Reinitialize a state object
}
