% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SCColumns}
\alias{SCColumns}
\title{Column names of the Science Cloud input file in the preferred order.}
\usage{
SCColumns()
}
\value{
a character vector of the row Science Cloud column names in the order they
should be displayed in the GUI and the report excel sheet "Science Cloud Input"
}
\description{
Column names of the Science Cloud input file in the preferred order.
}
