% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SetEvent}
\alias{SetEvent}
\title{Set event for input parameter id in state}
\usage{
SetEvent(state, id, value)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}

\item{value}{character string, one of "USER", "SCDATA", "REPORT", "INIT", "DEFAULT", "MIN", "MAX".}
}
\value{
nothing - this function only has side effects: setting the \code{state$event[[id]]}.
}
\description{
Set event for input parameter id in state
}
\details{
This function is supposed to be called before calling an observer code would call Process...Event functions.
}
