% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SetReportInput}
\alias{SetReportInput}
\title{Set the value for an input parameter id read from a report}
\usage{
SetReportInput(
  state,
  id,
  value,
  flagSynchronous = getOption("MMVshiny.synchSetReportInput", FALSE)
)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}

\item{value}{the value to set for id in state.}

\item{flagSynchronous}{a logical indicating whether the processing of the input, including validation
and setting the status, should be synchronous. Default is \code{getOption("MMVshiny.synchSetReportInput", FALSE)}.
Beaware that setting this to TRUE can lead to unexpected behavior if you have an observer on
GetReportInput() for the same id.}
}
\value{
nothing - this function only has side effects: setting the \code{state$reportInput[[id]]}.
}
\description{
Set the value for an input parameter id read from a report
}
\details{
This function should be called within a reactive context.
}
