% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SetSCInput}
\alias{SetSCInput}
\title{Set the value for an id averaged from an uploaded Science Cloud file}
\usage{
SetSCInput(
  state,
  id,
  value,
  flagSynchronous = getOption("MMVshiny.synchSetSCInput", FALSE)
)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of the input parameter}

\item{value}{the value to set}

\item{flagSynchronous}{logical indicating if the SCDATA event should be
processed within this function call (TRUE) or if it should be processed when an observer
on GetSCInput(id) is triggered (FALSE). By default this is set to
\code{getOption("MMVshiny.synchSetSCInput", FALSE)}.
Pay attention that setting this argument to TRUE may lead to unexpected behavior if there
is an observer on GetSCInput(id).}
}
\description{
Set the value for an id averaged from an uploaded Science Cloud file
}
