% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SetSCRawData}
\alias{SetSCRawData}
\title{Set Science Cloud data in the state object}
\usage{
SetSCRawData(
  state,
  scRawData,
  flagSetEvent = TRUE,
  flagSynchronous = getOption("MMVshiny.synchSetSCInput", FALSE)
)
}
\arguments{
\item{state}{a state object.}

\item{scRawData}{a data.frame of data from Science Cloud excel sheet}

\item{flagSetEvent}{logical (default: TRUE) indicating if an SCDATA event should be
fired. Currently, the purpose of this argument is to prevent firing SCDATA events
when loading Science Cloud raw data from a MMVSola report file.}

\item{flagSynchronous}{logical indicating if the SCDATA event should be
processed within this function call (TRUE) or if it should be processed when an observer
on GetSCInput(id) is triggered (FALSE). By default this is set to \code{getOption("MMVshiny.synchSetSCInput", FALSE)}.
Pay attention that setting this argument to TRUE may lead to unexpected behavior if there
is an observer on GetSCInput(id). This argument is passed to SetSCInput.}
}
\value{
nothing. This function only has side effects.
}
\description{
Set Science Cloud data in the state object
}
