% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{SetStatus}
\alias{SetStatus}
\title{Set the status for an input parameter id in a state object}
\usage{
SetStatus(state, id, ...)
}
\arguments{
\item{state}{the state object.}

\item{id}{the id of the input parameter.}

\item{...}{a name-value pairs to update the status for id in state.}
}
\value{
nothing - this function only has side effects: updating the status for id in state.
}
\description{
Set the status for an input parameter id in a state object
}
\details{
Existing name value-pairs in the status for id in state are updated with the new values or
kept unchanged if not specified in the ... argument.
}
