% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{UpdateCheckboxInput}
\alias{UpdateCheckboxInput}
\title{Wrapper of shiny::updateCheckboxInput that also calls session$setInputs if session is a MockShinySession.}
\usage{
UpdateCheckboxInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}
}
\value{
result of shiny::updateCheckboxInput.
}
\description{
Wrapper of shiny::updateCheckboxInput that also calls session$setInputs if session is a MockShinySession.
}
\details{
This function is useful for testing shiny applications with shiny::testServer.
}
