% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{UpdateNumericInput}
\alias{UpdateNumericInput}
\title{Wrapper of shiny::updateNumericInput that also calls session$setInputs if session is a MockShinySession.}
\usage{
UpdateNumericInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}
}
\value{
result of shiny::updateNumericInput.
}
\description{
Wrapper of shiny::updateNumericInput that also calls session$setInputs if session is a MockShinySession.
}
\details{
This function is useful for testing shiny applications with shiny::testSever.
}
