% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{UpdateTextInput}
\alias{UpdateTextInput}
\title{Wrapper of shiny::updateTextInput that also calls session$setInputs if session is a MockShinySession.}
\usage{
UpdateTextInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\value{
result of shiny::updateTextInput.
}
\description{
Wrapper of shiny::updateTextInput that also calls session$setInputs if session is a MockShinySession.
}
\details{
This function is useful for testing shiny applications with shiny::testSever.
}
