% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{Validate}
\alias{Validate}
\title{Set the validated state for input parameter id, based on the available report input, gui input,
dependancy input values and default value}
\usage{
Validate(state, id, logMessage)
}
\arguments{
\item{state}{the state object.}

\item{id}{id of the input parameter.}

\item{logMessage}{a character string used for debugging purpose only.}
}
\value{
the validated value. This function has side effects on the state object - it updates
the validated value and the status for id.
}
\description{
Set the validated state for input parameter id, based on the available report input, gui input,
dependancy input values and default value
}
