% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ValidateBoundaries}
\alias{ValidateBoundaries}
\title{Validate that a numeric input value is within \verb{[min,max]} interval}
\usage{
ValidateBoundaries(state, id, value)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a numeric input parameter.}

\item{value}{a numeric vector containing the input values to
validate. Each value of this vector is checked to be within the interval
\verb{[GetMin(state, id),GetMax(state,id)]} and if it is outside the interval,
this value is replaced depending on \code{GetOutOfBounds(state,id)}: if 'nearest',
the nearest boudnary is used; if 'na', NA_real_ is used.
The vector can be of length bigger than one, in particular,
if the same numeric input is informed from a Science Cloud file.}
}
\value{
a numeric vector of the same length as \code{value}. As a side effect, SetStatus is called on id
with setting validationNote to a text string.
}
\description{
Validate that a numeric input value is within \verb{[min,max]} interval
}
