% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ValidateText}
\alias{ValidateText}
\title{Validate a text input value}
\usage{
ValidateText(state, id, value)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a text input parameter.}

\item{value}{a character string or a vector of such.}
}
\value{
a character string.
}
\description{
Validate a text input value
}
