% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{CreateUIInput}
\alias{CreateUIInput}
\title{Create an input field and, optionally, a status icon according to the type of id in a spec}
\usage{
CreateUIInput(
  state,
  id,
  spec = state$spec,
  useLabel = TRUE,
  iconValign = "middle",
  ...
)
}
\arguments{
\item{state}{the state object. Can be missing, in which case, the function assumes that the
create UI element is global for all states, so that the id of the element is equal to the argument
\code{id}. If state is provided, the id of the created element will be \code{GetGuiId(state, id)}.}

\item{id}{the id of the input parameter for which an input field will be created.}

\item{spec}{the specification object. Default is \code{state$spec}.}

\item{useLabel}{logical indicating whether to use the GUILABEL from the spec. Default is TRUE.}

\item{iconValign}{the vertical alignment of the status icon. Default is "middle". This argument is only
used if \code{spec[ID == id, STATUSICON == TRUE]}.}

\item{...}{additional arguments to be passed to the input function.}
}
\value{
a list of td elements: one for the input and, another one for the status icon. If
\code{spec[ID == id, STATUSICON == FALSE]}, only a single td element is returned.
}
\description{
Create an input field and, optionally, a status icon according to the type of id in a spec
}
