% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GenerateJavaScriptEventHandlers}
\alias{GenerateJavaScriptEventHandlers}
\title{Generate JS code for event handlers on the client's browser side}
\usage{
GenerateJavaScriptEventHandlers(
  spec,
  ids = spec[grepl("input", TYPE), ID],
  filename
)
}
\arguments{
\item{spec}{parameter specification.}

\item{ids}{a character vector of ids for which event handlers should be generated. Default
is all input parameters in spec, i.e. \code{spec[grepl("input", TYPE), ID]}.}

\item{filename}{a character string path to an R-file where the R-code will be generated. If
not specified a tempfile is created.}
}
\value{
a character string - the filepath to the generated R-script (same as filename if this
argument was provided).
}
\description{
Generate JS code for event handlers on the client's browser side
}
