% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GenerateScriptCreatingObservers}
\alias{GenerateScriptCreatingObservers}
\title{Generate observeEvent calls and store them in an R-file to be sourced from within shiny server or state environment}
\usage{
GenerateScriptCreatingObservers(
  spec,
  inputObjectName = "input",
  outputObjectName = "output",
  stateObjectName = "state",
  ids = unique(spec$ID),
  observerTypes = c("countID", "default", "displayed", "SCInput", "reportInput",
    "resetCount", "min", "max"),
  filename
)
}
\arguments{
\item{spec}{parameter specification.}

\item{inputObjectName}{a character string indicating the name of the shiny input object (default "input").}

\item{outputObjectName}{a character string indicating the name of the shiny output object (default "output").}

\item{stateObjectName}{a character string indicating the name of the MMVSola state object (default: "state").}

\item{ids}{a character vector of ids for which observers should be generated.}

\item{observerTypes}{a character vector of observer types to be generated. Possible values
are "countID", "default", "displayed", "SCInput", "reportInput", "resetCount", "min", "max". Bye default, this
is a vector of all of these. Some observer types may not be relevant for some input types, for example,
"min" and "max" are only relevant for numeric inputs, and countID and displayed are only relevant for inputs
with a GUI label.}

\item{filename}{a character string path to an R-file where the R-code will be generated. If
not specified a tempfile is created.}
}
\value{
a character string - the filepath to the generated R-script (same as filename if this
argument was provided).
}
\description{
Generate observeEvent calls and store them in an R-file to be sourced from within shiny server or state environment
}
