% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GenerateScriptCreatingReactives}
\alias{GenerateScriptCreatingReactives}
\title{Generate an R script creating reactive objects inside the server function for each value in state}
\usage{
GenerateScriptCreatingReactives(
  spec,
  stateObjectName = "state",
  ids = spec[, unique(ID)],
  filename
)
}
\arguments{
\item{spec}{parameter specification table.}

\item{stateObjectName}{a character string indicating the name of the MMVSola state object (default: "state").}

\item{ids}{a character vector of ids for which reactive objects should be created. Default is all ids in spec.}

\item{filename}{a character string path to an R-file where the R-code will be generated. If
not specified a tempfile is created.}
}
\value{
a character string - the filepath to the generated R-script (same as filename if this
argument was provided).
}
\description{
Generate an R script creating reactive objects inside the server function for each value in state
}
