% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GenerateScriptRenderingIcons}
\alias{GenerateScriptRenderingIcons}
\title{Generate a R script calling renderUI for each status icon in a spec}
\usage{
GenerateScriptRenderingIcons(
  spec,
  prefixGuiId = "",
  outputObjectName = "output",
  stateObjectName = "state",
  ids = spec[STATUSICON == TRUE, ID],
  filename
)
}
\arguments{
\item{spec}{parameter specification table.}

\item{prefixGuiId}{a character string indicating the prefix for the gui id (default "").}

\item{outputObjectName}{a character string indicating the name of the shiny output object (default "output").}

\item{stateObjectName}{a character string indicating the name of the MMVSola state object (default: "state").}

\item{ids}{a character vector of ids for which status icons should be rendered. Default is all ids in spec with
STATUSICON == TRUE.}

\item{filename}{a character string path to an R-file where the R-code will be generated. If
not specified a tempfile is created.}
}
\value{
a character string - the filepath to the generated R-script (same as filename if this
argument was provided).
}
\description{
Generate a R script calling renderUI for each status icon in a spec
}
