% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{GetReportLabel}
\alias{GetReportLabel}
\title{Get the REPORTLABEL for one or several ID's in the spec}
\usage{
GetReportLabel(state, ids, spec = state$spec)
}
\arguments{
\item{state}{a state object.}

\item{ids}{a character vector indicating ids for which the REPORTLABEL should be returned.}

\item{spec}{the parameter specification table. Default: \code{state$spec}.}
}
\value{
a character vector
}
\description{
Get the REPORTLABEL for one or several ID's in the spec
}
