% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{NAVal}
\alias{NAVal}
\title{Get the NA value of the correct type corresponding to id}
\usage{
NAVal(state, id, spec = state$spec)
}
\arguments{
\item{state}{the state object}

\item{id}{the id of the input}

\item{spec}{the spec object. Default is state$spec.}
}
\value{
the NA value of the correct type
}
\description{
Get the NA value of the correct type corresponding to id
}
