% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{UpdateSelectInput}
\alias{UpdateSelectInput}
\title{Wrapper of shiny::updateSelectInput that also calls session$setInputs if session is a MockShinySession.}
\usage{
UpdateSelectInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{choices}{List of values to select from. If elements of the list are
named, then that name --- rather than the value --- is displayed to the
user. It's also possible to group related inputs by providing a named list
whose elements are (either named or unnamed) lists, vectors, or factors. In
this case, the outermost names will be used as the group labels (leveraging
the \verb{<optgroup>} HTML tag) for the elements in the respective sublist. See
the example section for a small demo of this feature.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). If not specified then defaults to the first value for
single-select lists and no values for multiple select lists.}
}
\value{
result of shiny::updateSelectInput.
}
\description{
Wrapper of shiny::updateSelectInput that also calls session$setInputs if session is a MockShinySession.
}
\details{
This function is useful for testing shiny applications with shiny::testSever.
}
