% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ValidateRadio}
\alias{ValidateRadio}
\title{Validate a radio input value}
\usage{
ValidateRadio(state, id, value)
}
\arguments{
\item{state}{a state object}

\item{id}{the id of a radio input parameter.}

\item{value}{a character vector of length 1.}
}
\value{
if value is within the radio choices for id, the same value is returned, otherwise an error is raised.
}
\description{
Validate a radio input value
}
