% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessState.R
\name{ValidateValue}
\alias{ValidateValue}
\title{Get the validated value for id given a suggested value, without updating
the validated state for id}
\usage{
ValidateValue(state, id, value)
}
\arguments{
\item{state}{a state object.}

\item{id}{a character string id.}

\item{value}{the suggested value to validate.}
}
\value{
the validated value corresponding to value
}
\description{
Get the validated value for id given a suggested value, without updating
the validated state for id
}
